requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Tokenization
tokenizers>=0.13.0
sentencepiece>=0.1.99

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
PyPDF2>=3.0.0
pdf2image>=1.16.0

# OCR
pytesseract>=0.3.10
opencv-python>=4.8.0

# Training & Monitoring
wandb>=0.15.0
tqdm>=4.65.0
matplotlib>=3.7.0

# API & Deployment
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.31.0
